# src/main/resources/swagger.yaml
swagger: "2.0"
info:
  description: "Maantrack."
  version: "1.0.0"
  title: "Maantrack"
  termsOfService: ""
  contact:
    email: "pravinrahanagdale41@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/"
...
schemes:
  - "http"
  - "https"
paths:
...
  /user/{userId}:
    get:
      tags:
        - "user"
      summary: "Get user by user id"
      description: ""
      operationId: "getUserById"
      security:
        - api_key: []
      produces:
        - "application/json"
      parameters:
        - name: "userId"
          in: "path"
          description: "The user id that needs to be fetched. Use 1 for testing. "
          required: true
          type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid username supplied"
        404:
          description: "User not found"
    delete:
      security:
        - api_key: []
      tags:
        - "user"
      summary: "Delete user"
      description: "This can only be done by the logged in user."
      operationId: "deleteUser"
      produces:
        - "application/json"
      parameters:
        - name: "userId"
          in: "path"
          description: "The user id that needs to be deleted"
          required: true
          type: "string"
      responses:
        400:
          description: "Invalid username supplied"
        404:
          description: "User not found"
...
securityDefinitions:
  api_key:
    type: "apiKey"
    in: "header"
    name: "authorization"
definitions:
...
  User:
    type: "object"
    required:
      - "username"
    properties:
      usersId:
        type: "integer"
        format: "int64"
      avatarUrl:
        type: "string"
      avatarSource:
        type: "string"
      bio:
        type: "string"
      confirmed:
        type: "string"
      email:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      userType:
        type: "string"
        enum: [User, Administrator]
      profileUrl:
        type: "string"
      password:
        type: "string"
      userName:
        type: "string"
      birthDate:
        type: "string"
        format: "date-time"
      createdDate:
        type: "string"
        format: "date-time"
      modifiedDate:
        type: "string"
        format: "date-time"
    xml:
      name: "User"
...